requirements.txt
certifi==2018.8.13
chardet==3.0.4
Django==2.1
django-environ==0.4.5
djangorestframework==3.8.2
environ==1.0
httpie==0.9.9
idna==2.7
mysqlclient==1.3.13
Pygments==2.2.0
pytz==2018.5
requests==2.19.1
urllib3==1.23
